+++ CHANGELOG.md
@@ -1,7 +1,14 @@
+#### [0.8.8] - Nov 21, 2023
+
+- add `onTapOutside` callback to SearchField [Issue #94](https://github.com/maheshmnj/searchfield/issues/94)
+- add `autofocus` property to SearchField [Issue #105](https://github.com/maheshmnj/searchfield/issues/105)
+
 #### [0.8.7] - Oct 27, 2023
+
 - Minor Fix in suggestions total height calculation. [#PR 85](https://github.com/maheshmnj/searchfield/pull/85)
 
 #### [0.8.6] - Oct 13, 2023
+
 - Remove default scrollbar from listview
 
 #### [0.8.5] - Oct 12, 2023